libs/python/mylibrary/pyproject.toml
[project]
name = "mylibrary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "yurikrupnik", email = "krupnik.yuri@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
